Refactor DocDesc and DLPlate definition to DocumentLanguage.Core.
