+++ code/drasil-docLang/Drasil/ExtractDocDesc.hs
@@ -2,100 +2,14 @@
 module Drasil.ExtractDocDesc (getDocDesc, egetDocDesc, ciGetDocDesc) where
 
 import Control.Lens((^.))
-import Drasil.DocumentLanguage
+import Drasil.DocumentLanguage.Core
 import Language.Drasil hiding (Manual, Vector, Verb)
 import Theory.Drasil (Theory(..))
 import Data.List(transpose)
 
 import Data.Functor.Constant (Constant(Constant))
-import Data.Generics.Multiplate (Multiplate(multiplate, mkPlate), appendPlate,
+import Data.Generics.Multiplate (appendPlate,
   foldFor, purePlate, preorderFold)
-
-data DLPlate f = DLPlate {
-  docSec :: DocSection -> f DocSection,
-  refSec :: RefSec -> f RefSec,
-  introSec :: IntroSec -> f IntroSec,
-  introSub :: IntroSub -> f IntroSub,
-  stkSec :: StkhldrSec -> f StkhldrSec,
-  stkSub :: StkhldrSub -> f StkhldrSub,
-  gsdSec :: GSDSec -> f GSDSec,
-  gsdSub :: GSDSub -> f GSDSub,
-  ssdSec :: SSDSec -> f SSDSec,
-  ssdSub :: SSDSub -> f SSDSub,
-  pdSec :: ProblemDescription -> f ProblemDescription,
-  pdSub :: PDSub -> f PDSub,
-  scsSub :: SCSSub -> f SCSSub,
-  reqSec :: ReqrmntSec -> f ReqrmntSec,
-  reqSub :: ReqsSub -> f ReqsSub,
-  lcsSec :: LCsSec -> f LCsSec,
-  ucsSec :: UCsSec -> f UCsSec,
-  traceSec :: TraceabilitySec -> f TraceabilitySec,
-  offShelfSolnsSec :: OffShelfSolnsSec -> f OffShelfSolnsSec,
-  auxConsSec :: AuxConstntSec -> f AuxConstntSec,
-  appendSec :: AppndxSec -> f AppndxSec
-}
-
-instance Multiplate DLPlate where
-  multiplate p = DLPlate ds res intro intro' stk stk' gs gs' ss ss' pd pd' sc
-    rs rs' lcp ucp ts es acs aps where
-    ds (RefSec x) = RefSec <$> refSec p x
-    ds (IntroSec x) = IntroSec <$> introSec p x
-    ds (StkhldrSec x) = StkhldrSec <$> stkSec p x
-    ds (GSDSec x) = GSDSec <$> gsdSec p x
-    ds (SSDSec x) = SSDSec <$> ssdSec p x
-    ds (ReqrmntSec x) = ReqrmntSec <$> reqSec p x
-    ds (LCsSec x) = LCsSec <$> lcsSec p x
-    ds (UCsSec x) = UCsSec <$> ucsSec p x
-    ds (TraceabilitySec x) = TraceabilitySec <$> traceSec p x
-    ds (OffShelfSolnsSec x) = OffShelfSolnsSec <$> offShelfSolnsSec p x
-    ds (AuxConstntSec x) = AuxConstntSec <$> auxConsSec p x
-    ds (AppndxSec x) = AppndxSec <$> appendSec p x
-    ds Bibliography = pure Bibliography
-
-    res (RefProg c x) = RefProg <$> pure c <*> pure x
-    intro (IntroProg s1 s2 progs) = IntroProg <$> pure s1 <*> pure s2 <*>
-      traverse (introSub p) progs
-    intro' (IPurpose s) = IPurpose <$> pure s
-    intro' (IScope s1 s2) = IScope <$> pure s1 <*> pure s2
-    intro' (IChar s1 s2 s3) = IChar <$> pure s1 <*> pure s2 <*> pure s3
-    intro' (IOrgSec s1 c sect s2) = IOrgSec <$> pure s1 <*> pure c <*> pure sect <*> pure s2
-    stk (StkhldrProg c s) = StkhldrProg <$> pure c <*> pure s
-    stk (StkhldrProg2 progs) = StkhldrProg2 <$> traverse (stkSub p) progs
-    stk' (Client c s) = Client <$> pure c <*> pure s
-    stk' (Cstmr c) = Cstmr <$> pure c
-    gs (GSDProg s1 c labcon s2) = GSDProg <$> pure s1 <*> pure c <*> pure labcon <*> pure s2
-    gs (GSDProg2 x) = GSDProg2 <$> traverse (gsdSub p) x
-    gs' (SysCntxt c) = SysCntxt <$> pure c
-    gs' (UsrChars c) = UsrChars <$> pure c
-    gs' (SystCons c s) = SystCons <$> pure c <*> pure s
-    ss (SSDProg progs) = SSDProg <$> traverse (ssdSub p) progs
-    ss' (SSDProblem prog) = SSDProblem <$> pdSec p prog
-    ss' (SSDSolChSpec (SCSProg spec)) = SSDSolChSpec . SCSProg <$> traverse (scsSub p) spec
-    pd (PDProg s sect progs) = PDProg <$> pure s <*> pure sect <*> traverse (pdSub p) progs
-    pd' (TermsAndDefs s cs) = TermsAndDefs <$> pure s <*> pure cs
-    pd' (Goals s ci) = Goals <$> pure s <*> pure ci
-    pd' (PhySysDesc nm s lc c) = PhySysDesc <$> pure nm <*> pure s <*> pure lc <*> pure c
-    sc Assumptions = pure Assumptions
-    sc (TMs s f t) = TMs <$> pure s <*> pure f <*> pure t
-    sc (GDs s f g d) = GDs <$> pure s <*> pure f <*> pure g <*> pure d
-    sc (DDs s f dd d) = DDs <$> pure s <*> pure f <*> pure dd <*> pure d
-    sc (IMs s f i d) = IMs <$> pure s <*> pure f <*> pure i <*> pure d
-    sc (Constraints s1 s2 s3 l) = Constraints <$> pure s1 <*> pure s2 <*> pure s3 <*> pure l
-    sc (CorrSolnPpties c) = CorrSolnPpties <$> pure c
-    rs (ReqsProg reqs) = ReqsProg <$> traverse (reqSub p) reqs
-    rs' (FReqsSub ci con) = FReqsSub <$> pure ci <*> pure con
-    rs' (NonFReqsSub c) = NonFReqsSub <$> pure c
-    lcp (LCsProg c) = LCsProg <$> pure c
-    ucp (UCsProg c) = UCsProg <$> pure c
-    ts (TraceabilityProg llc sen con sect) = TraceabilityProg <$> pure llc <*>
-      pure sen <*> pure con <*> pure sect
-    es (OffShelfSolnsProg contents) = OffShelfSolnsProg <$> pure contents
-    acs (AuxConsProg ci qdef) = AuxConsProg <$> pure ci <*> pure qdef
-    aps (AppndxProg con) = AppndxProg <$> pure con
-  mkPlate b = DLPlate (b docSec) (b refSec) (b introSec) (b introSub) (b stkSec)
-    (b stkSub) (b gsdSec) (b gsdSub) (b ssdSec) (b ssdSub) (b pdSec) (b pdSub)
-    (b scsSub) (b reqSec) (b reqSub) (b lcsSec) (b ucsSec)
-    (b traceSec) (b offShelfSolnsSec) (b auxConsSec) (b appendSec)
 
 secConPlate :: Monoid b => (forall a. HasContents a => [a] -> b) ->
   ([Section] -> b) -> DLPlate (Constant b)
@@ -127,7 +41,7 @@ secConPlate mCon mSec = preorderFold $ purePlate {
   ucsSec = Constant <$> \(UCsProg _) -> mempty,
   traceSec = Constant <$> \(TraceabilityProg lc _ c s) ->
     mconcat [mCon lc, mCon c, mSec s],
-  offShelfSolnsSec = Constant <$> \(OffShelfSolnsProg c) -> mCon c,
+  offShelfSec = Constant <$> \(OffShelfSolnsProg c) -> mCon c,
   appendSec = Constant <$> \(AppndxProg c) -> mCon c
 }
