+++ code/drasil-docLang/Drasil/TraceTable.hs
@@ -16,7 +16,7 @@ import Language.Drasil.Development (lnames')
 import Database.Drasil (TraceMap, traceMap)
 import Theory.Drasil (DataDefinition, GenDefn, InstanceModel, TheoryModel)
 
-import Drasil.DocumentLanguage
+import Drasil.DocumentLanguage.Core
 
 
 traceMap' :: HasUID l => (l -> [Sentence]) -> [l] -> TraceMap
